# backend/internal/http/http.go
// Package http 是 mdm-product 的 REST 面（路径前缀 /mdm/product，与
// assembly.yaml 的 edge_routes 一致）。/healthz、/metrics 由
// besdk.NewGinEngine 挂好（/healthz 只答进程活着、不查依赖），这里不重复挂，
// 也不写进 openapi（契约以 /mdm/product 为前缀，写进去会暗示错误的路径）。
package http

import (
	"fmt"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"

	besdk "github.com/brickKit/be-sdk-go"
	"github.com/brickKit/mdm-product/v2/backend/internal/repo"
	"github.com/brickKit/mdm-product/v2/backend/internal/service"
)

// RegisterRoutes 挂载业务路由。每条路由都带 assembly.yaml 里声明的权限键
// 注册（besdk.GET / POST / PATCH），没有不鉴权的业务路由。
//
// 启用 / 停用单独用 mdm.product.set_status，不跟编辑共用 mdm.product.edit：
// 停用一个产品会让它从所有选择器里消失，"能改名字就能停用"给的权限太宽。
// convert-quantity 是不落库的试算，归 view。data_scopes: none：产品主数据
// 全员可见，不做行级过滤。
func RegisterRoutes(eng *gin.Engine, svc *service.Service) {
	g := eng.Group("/mdm/product")
	besdk.GET(g, "/products", "mdm.product.view", listHandler(svc))
	besdk.GET(g, "/products/:id", "mdm.product.view", getHandler(svc))
	besdk.POST(g, "/products", "mdm.product.create", createHandler(svc))
	besdk.PATCH(g, "/products/:id", "mdm.product.edit", updateHandler(svc))
	besdk.POST(g, "/products/:id/status", "mdm.product.set_status", setStatusHandler(svc))
	besdk.POST(g, "/products/convert-quantity", "mdm.product.view", convertQuantityHandler(svc))
}

type productDTO struct {
	ID           string `json:"id"`
	SKU          string `json:"sku"`
	Name         string `json:"name"`
	CategoryID   string `json:"category_id"`
	BaseUOMID    string `json:"base_uom_id"`
	TrackingType string `json:"tracking_type"`
	StandardCost string `json:"standard_cost"`
	Status       string `json:"status"`
	Version      int64  `json:"version"`
	CreatedAt    string `json:"created_at"`
	UpdatedAt    string `json:"updated_at"`
}

func toDTO(p *repo.Product) productDTO {
	return productDTO{
		ID: p.ID, SKU: p.SKU, Name: p.Name, CategoryID: p.CategoryID,
		BaseUOMID: p.BaseUOMID, TrackingType: p.TrackingType, StandardCost: p.StandardCost,
		Status: p.Status, Version: p.Version,
		CreatedAt: p.CreatedAt.Format(rfc3339), UpdatedAt: p.UpdatedAt.Format(rfc3339),
	}
}

const rfc3339 = "2006-01-02T15:04:05.999999999Z07:00"

type createRequest struct {
	IdempotencyKey string `json:"idempotency_key" binding:"required"`
	SKU            string `json:"sku"`
	Name           string `json:"name" binding:"required"`
	CategoryID     string `json:"category_id"`
	BaseUOMID      string `json:"base_uom_id" binding:"required"`
	TrackingType   string `json:"tracking_type"`
	StandardCost   string `json:"standard_cost"`
}

func createHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req createRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		out, err := svc.Create(c.Request.Context(), repo.CreateInput{
			IdempotencyKey: req.IdempotencyKey, SKU: req.SKU, Name: req.Name,
			CategoryID: req.CategoryID, BaseUOMID: req.BaseUOMID,
			TrackingType: req.TrackingType, StandardCost: req.StandardCost,
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, toDTO(out))
	}
}

func getHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		out, err := svc.Get(c.Request.Context(), c.Param("id"))
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, toDTO(out))
	}
}

func listHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		in, err := parseListInput(c)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		out, err := svc.List(c.Request.Context(), in)
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		dtos := make([]productDTO, 0, len(out.Products))
		for _, p := range out.Products {
			dtos = append(dtos, toDTO(p))
		}
		c.JSON(http.StatusOK, gin.H{"products": dtos, "next_cursor": out.NextCursor})
	}
}

// parseListInput 读 GET /products 的查询参数。created_after / created_before
// 是 RFC 3339 时间，格式不对就是调用方的错（400），不悄悄退回默认窗口。
// page_size 不是数字时按 0 处理，由 besdk.ListWindow 取默认页大小。
func parseListInput(c *gin.Context) (repo.ListInput, error) {
	pageSize, _ := strconv.Atoi(c.Query("page_size"))
	in := repo.ListInput{
		Q:            c.Query("q"),
		Cursor:       c.Query("cursor"),
		PageSize:     pageSize,
		StatusFilter: c.Query("status_filter"),
	}
	for _, f := range []struct {
		name string
		dst  *time.Time
	}{{"created_after", &in.CreatedAfter}, {"created_before", &in.CreatedBefore}} {
		raw := c.Query(f.name)
		if raw == "" {
			continue
		}
		t, err := time.Parse(time.RFC3339, raw)
		if err != nil {
			return repo.ListInput{}, fmt.Errorf("%s 不是 RFC 3339 时间：%q", f.name, raw)
		}
		*f.dst = t
	}
	return in, nil
}

type updateRequest struct {
	IdempotencyKey string `json:"idempotency_key" binding:"required"`
	Version        int64  `json:"version" binding:"required"`
	Name           string `json:"name"`
	CategoryID     string `json:"category_id"`
	StandardCost   string `json:"standard_cost"`
}

func updateHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req updateRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		out, err := svc.Update(c.Request.Context(), service.UpdateInput{
			IdempotencyKey: req.IdempotencyKey, ID: c.Param("id"), Version: req.Version,
			Name: req.Name, CategoryID: req.CategoryID, StandardCost: req.StandardCost,
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, toDTO(out))
	}
}

type setStatusRequest struct {
	IdempotencyKey string `json:"idempotency_key" binding:"required"`
	Version        int64  `json:"version" binding:"required"`
	Status         string `json:"status" binding:"required"`
}

func setStatusHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req setStatusRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		out, err := svc.SetStatus(c.Request.Context(), service.SetStatusInput{
			IdempotencyKey: req.IdempotencyKey, ID: c.Param("id"),
			Version: req.Version, Status: req.Status,
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, toDTO(out))
	}
}

type convertQuantityRequest struct {
	ProductID string `json:"product_id" binding:"required"`
	Qty       string `json:"qty" binding:"required"`
	FromUOMID string `json:"from_uom_id" binding:"required"`
	ToUOMID   string `json:"to_uom_id" binding:"required"`
}

func convertQuantityHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req convertQuantityRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		qty, err := svc.ConvertQuantity(c.Request.Context(), req.ProductID, req.Qty, req.FromUOMID, req.ToUOMID)
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, gin.H{"qty": qty})
	}
}