# contracts/product.openapi.yaml
openapi: 3.0.3
info:
  title: mdm-product · 产品/物料主数据
  version: 2.0.0
  description: |
    对外 REST 面；别的组件走 gRPC（contracts/mdm/product/v1/product.proto）。
    BatchGet / GetSummary 只在 gRPC 上：它们是给别的组件防 N+1 的批量读，
    暴露成公网 REST 等于给了一个"按 id 列表一次拿走一批产品"的接口。
    ConvertQuantity 例外：前端下单页要实时试算"3 箱 = 36 个"，换算规则只在
    后端，前端不自己算，所以它在 REST 上。

    /healthz 不在这份契约里：它是 be-sdk-go 的 NewGinEngine 给每个组件统一
    挂的裸根路径，不带 /mdm/product 前缀。

servers:
  - url: /mdm/product

paths:
  /products:
    get:
      summary: List
      operationId: listProducts
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: |
            关键字，匹配 sku 或 name，大小写不敏感，% 与 _ 按字面匹配。sku 或
            name 以 q 开头的排在只是包含 q 的前面，同一档内按建档时间倒序。去掉
            首尾空白后为空等于不传。时间窗口照常生效：要搜更早建档的产品，显式
            给 created_after。
        - name: cursor
          in: query
          schema: { type: string }
          description: 上一页返回的 next_cursor，原样回传；只对发出它的那组查询参数有效。没有 offset 参数。
        - name: page_size
          in: query
          schema: { type: integer }
        - name: status_filter
          in: query
          schema: { $ref: "#/components/schemas/ProductStatus" }
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
          description: 不传时间范围时只查最近 90 天建档的产品。格式不是 RFC 3339 时回 400。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
                  next_cursor: { type: string }
    post:
      summary: Create
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }

  /products/{id}:
    get:
      summary: Get
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    patch:
      summary: Update
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }

  /products/{id}/status:
    post:
      summary: SetStatus
      operationId: setProductStatus
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetStatusRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }

  /products/convert-quantity:
    post:
      summary: ConvertQuantity（DryRun，纯函数不落库）
      operationId: convertProductQuantity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConvertQuantityRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConvertQuantityResponse" }

components:
  schemas:
    ProductStatus:
      type: string
      enum: [ACTIVE, DISABLED]

    TrackingType:
      type: string
      enum: [NONE, BATCH, SERIAL]

    Product:
      type: object
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        category_id: { type: string }
        base_uom_id: { type: string }
        tracking_type: { $ref: "#/components/schemas/TrackingType" }
        standard_cost:
          type: string
          description: 金额一律用 string 传 decimal，绝不用 double（跨语言精度）
        status: { $ref: "#/components/schemas/ProductStatus" }
        version: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CreateRequest:
      type: object
      required: [idempotency_key, name, base_uom_id]
      properties:
        idempotency_key:
          type: string
          description: 跨重试的幂等键；同一个键重试返回第一次的结果，不重复执行
        sku:
          type: string
          description: 可选，留空则按 "P" + 6 位自增数字生成
        name: { type: string }
        category_id: { type: string }
        base_uom_id: { type: string }
        tracking_type: { $ref: "#/components/schemas/TrackingType" }
        standard_cost: { type: string }

    UpdateRequest:
      type: object
      required: [idempotency_key, version]
      properties:
        idempotency_key: { type: string }
        version:
          type: integer
          format: int64
          description: 乐观锁，与库里不一致则拒绝
        name: { type: string }
        category_id: { type: string }
        standard_cost: { type: string }

    SetStatusRequest:
      type: object
      required: [idempotency_key, version, status]
      properties:
        idempotency_key: { type: string }
        version: { type: integer, format: int64 }
        status: { $ref: "#/components/schemas/ProductStatus" }

    ConvertQuantityRequest:
      type: object
      required: [product_id, qty, from_uom_id, to_uom_id]
      properties:
        product_id: { type: string }
        qty: { type: string, description: decimal-as-string，同金额字段的理由 }
        from_uom_id: { type: string }
        to_uom_id: { type: string }

    ConvertQuantityResponse:
      type: object
      properties:
        qty:
          type: string
          description: 已按目标单位的 rounding 向上取整（CEIL）：出库时宁可多备、不可少发